pyproject.toml
[tool.poetry]
name = "gwf"
version = "1.7.0"
description = "A flexible, pragmatic workflow tool."
authors = ["Dan Søndergaard <das@birc.au.dk>", "Thomas Mailund <mailund@birc.au.dk>"]
license = "GPL-3.0"
readme = "README.rst"
documentation = "https://gwf.app/"
repository = "https://github.com/gwforg/gwf"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
    "Topic :: System :: Distributed Computing",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

packages = [
    { include = "gwf", from = "src" },
]

include = ["CHANGELOG.rst", "CONTRIBUTORS.rst", "LICENSE"]

[tool.poetry.scripts]
gwf = "gwf.cli:main"

[tool.poetry.plugins]

[tool.poetry.plugins."gwf.backends"]
slurm = "gwf.backends.slurm:SlurmBackend"
sge = "gwf.backends.sge:SGEBackend"
local = "gwf.backends.local:LocalBackend"
testing = "gwf.backends.testing:TestingBackend"

[tool.poetry.plugins."gwf.plugins"]
init = "gwf.plugins.init:init"
run = "gwf.plugins.run:run"
config = "gwf.plugins.config:config"
status = "gwf.plugins.status:status"
info = "gwf.plugins.info:info"
logs = "gwf.plugins.logs:logs"
clean = "gwf.plugins.clean:clean"
workers = "gwf.plugins.workers:workers"
cancel = "gwf.plugins.cancel:cancel"
touch = "gwf.plugins.touch:touch"

[tool.poetry.dependencies]
python = "^3.5"
click = "^6.0"
click-plugins = "^1.0"
lmdb = "^0.97.0"
attrs = "^19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-runner = "^5.1"
flake8 = "^3.7"
coverage = "^4.5"
coveralls = "^1.8"
sphinx = "^2.2"
sphinx-rtd-theme = "^0.4.3"
pytest-click = "^0.3.0"
pytest-mock = "^1.11"
pytest-cov = "^2.8"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
